refactor: reorganize modules for clarity
