+++ packages/text-field/package.json
@@ -19,6 +19,7 @@
   ],
   "dependencies": {
     "@hig/icon-button": "^0.1.3",
+    "@hig/utils": "^0.2.0",
     "classnames": "^2.2.5",
     "hig-react": "0.29.0",
     "prop-types": "^15.6.1",
