+++ packages/text-field/src/presenters/TextFieldPresenter.scss
@@ -1,7 +1,7 @@
 @import "~@hig/styles/mixins/colors.scss";
 @import "~@hig/styles/mixins/placeholder.scss";
 @import "~@hig/styles/mixins/typography.scss";
-@import "./tokens.scss";
+@import "../tokens.scss";
 
 @mixin styled-border($border-color, $hover-border-color: $border-color) {
   border: {
